+++ case/ppos2/test_locked_position_cofig.py
@@ -1,0 +1,45 @@
+# -*- coding: utf-8 -*-
+
+import json
+import math
+import time
+import random
+
+import allure
+import pytest
+from client_sdk_python import Web3
+from utils.platon_lib.ppos import Ppos
+from conf import  setting as conf
+from common import log
+import json
+from utils.platon_lib.ppos_common import CommonMethod
+from deploy.deploy import AutoDeployPlaton
+
+
+
+class TestLockeDposition:
+    address = conf.ADDRESS
+    pwd = conf.PASSWORD
+    abi = conf.DPOS_CONTRACT_ABI
+    cbft_json_path = conf.CBFT2
+    node_yml_path = conf.PPOS_NODE_TEST_YML
+    file = conf.CASE_DICT
+    privatekey = conf.PRIVATE_KEY
+    gasPrice = Web3.toWei(0.000000000000000001,'ether')
+    gas = 21000
+    transfer_gasPrice = Web3.toWei(1,'ether')
+    transfer_gas = 210000000
+    value = 1000
+    initial_amount = {'FOUNDATION': 905000000000000000000000000,
+                      'FOUNDATIONLOCKUP': 20000000000000000000000000,
+                      'STAKING': 25000000000000000000000000,
+                      'INCENTIVEPOOL': 45000000000000000000000000,
+                      'DEVELOPERS': 5000000000000000000000000
+                      }
+    def test(self):
+        CommonMethod.update_config('')
+        self.auto = AutoDeployPlaton ()
+        self.auto.start_all_node (self.node_yml_path)
+
+    def initial_unlock_Normal(self):
+
